Delete faulty ligand test
Why:
test_ligand_setup.py is breaking local tests.

Rather oddly, both nose and py.test produce
errors in TestImplicit and TestExplicit
test cases, but outside of any tests.

Somehow an import statement that fails
inside of this module gets called in this stage.
The error is something along the lines of
"ValueError: GAFF not found."

I'm deleting test_ligand_setup.py because
it is fauly and not necessary at the moment.
Once work on the ligand code resumes,
this code will need to be rewritten.

This change addresses the need by:

Deleting test_ligand_setup.py altogether.