Make ligand tests optional
Why:

The ligands module relies on OpenEye, but it is reasonable to
expect that not every user has access to OpenEye software.
We also currently don't have OpenEye set up
on our Travis builds (but we can fix this in the future).
It is to the developers discretion to run tests of the ligands module
locally, to make sure they are not broken.

To prevent the tests from failing entirely in these cases,
this module should now test whether it is possible to import the ligands
module, but if that fails, it will skip testing it without causing an error.

This change addresses the need by:

Adding a global mark to skip the tests if the ligands module is not
successfully imported, by catching the ImportError.