Add bond definitions and tripeptide testsystem
Why:

To prevent errors such as
```
ValueError: No template found for residue 1 (ACE).
The set of atoms matches ACE, but the bonds are different.
Perhaps the chain is missing a terminal group?
```

Definitions are needed for the bonds present in non-standard residues.
These have now been added to the files.

To test the new bonds, we've added a test for the new file, and a new
test system to run it on.

This change addresses the need by:

* Adding the new `bonds-protons.xml` file, and the link in `ff.py`.
* Adding the script to generate it, called `4-create_bond_definitions.py`
* Updating forcefield readme accordingly.
* Adding the `tripeptides` testsystem, containing two pdb files of tripeptides.