Add tripeptide test system
Why:

To correctly recognize residues, bond definitions are needed
for the custom amino acids that are part of the amber force field.
To test that the force field files are correctly formatted, and
can be used to set up a system, a tripeptide test system was added
to the testystems.

This change addresses the need by:

* Adding glu-ala-his and his-ala-his peptide files for use as
test systems.
* Add tests of the glu-ala-his system, to ensure protons can read the files
and set up a driver correctly.