Add reference GBAOAB implementation
Why:

The openmmtools codebase seems to be undergoing severe API changes every
other day. In the interest of consistency and correctness, I've implemented
a reference implementation of GBAOAB integration method.

I dont even know what is going on with the integrators in openmmtools anymore.
We will use this for protons while the openmmtools development carries on
without considering the burden of maintaining compatibility with this codebase.

This change addresses the need by:

* Adding a ReferenceGBAOABIntegrator to protons.integrators.
* Adding tests for this integrator in protons/tests/test_integrators.py
* Updating the tests and code in other locations to support this integrator.