Fix writing of netCDF on unexpected termination
Why:

When the script gets shut down because of exceeding wall time,
as opposed to keyboard exit, data does not get written to disk properly.
This code change causes an earlier write to disc during each call of
record_all, and optionally to each smaller record step.

This change addresses the need by:

* adding sync call to record_all.
* add option to call sync to record_drive_data
* add option to call sync to record_state_data
* add option to call sync to record_ghmc_integrator_data