Fix major protocol work reset bug
Why:

A bug was present in the code where an if statement was being bypassed
without throwing an error. Apparently, the integrators retrieved from a
CompoundIntegrator are all of type CustomIntegrator. This means that
the previously present if statement does not detect any of the integrators,
so NCMC work is not being reset at the start of each NCMC protocol.
This bug has been introduced around a month ago, when definition of the
compound integrator was changed to taking place outside of the Drive object.

This change addresses the need by:

* Replace the if statement in driver.py with a try except clause that
can handle both GHMC and BAOAB style integrators, and that throws
an error if it can not figure out how to reset the protocol work..