Fix import_gk_values
Why:

The old functionality of import_gk_values was linked to the internal
setup of calibration systems. It did not handle the ffxml residues,
such as ligands. I've now added bookkeeping of residue types, so that
gk values for non-standard ligands can also be provided.

This change addresses the need by:

* adding a residue type specifier to add_titratable_group
* updating import_gk_values