Add residue pooling feature
Why:

This addition allows the user to select different pools of residues
from which one can randomly sample protonation states. This will be
helpful if for instance you want to sample the ligand separately from
the protein, or want to specifically sample the protonation state
of specific residues.

This change addresses the need by:

The base class has been extended with a `define_pools` method which
allows the user to specify which indices should be sampled.
The `update` method (and the underlying `attempt_state_change`) have
been updated to make use of the user defined pools when desired by
passing a keyword argument.

A test has been added to test_ff which tries out pools on a tripeptide.