Fix atom indices bug in ForceFieldProtonDrive
Why:

This fixes a major bug in the `ForceFieldProtonDrive`. The bug was caused
by the assumption that atoms, for which indices from the topology were
retrieved, were assumed to be in the same order as the forcefield
template.

This fix makes use of code from `simtk.openmm.app.forcefield`
to match residue template atoms to topology atoms. It then makes sure
that the topology atom indices for each atom in the template are
assigned correctly.

The `ForceFieldProtonDrive` requires forcefield information to correctly
match indices, so this resulted in a breaking API change.
`ForceFieldProtonDrive` now requires the user to provide the `forcefield`
argument to `__init__`.

This change addresses the need by:

* Adding atom index matching code to the template reading code in
  `ForceFieldProtonDrive`'s add_xml_titration_groups
* Updating the API of `ForceFieldProtonDrive`, to now require forcefield as an argument.
* Tests have been updated to use the new API
* Add a test for the imidazole test system that checks against system modifications.