Add support for recording SAMS calibration data
Why:

I was relying on pickles to store data from calibrations. This allows for
storing it all in the same file. It also provides a space for metadata
that could be useful for resuming a stopped calibration.

This change addresses the need by:

* Add option to include `SelfAdjustedMixtureSampling` group for storing
data for calibrations.
* Adds optional variables for storing metadata.
* Add tests for new feature.
* Made GHMCIntegrator an optional part of the netcdf file, since we're not
using this integrator by default anymore.