Correctly propagate group index in SAMS
Why:

A bug was detected inside the SelfAdjustedMixtureSampling class.
The index of the titration group defaults to 0 in many places for
convenience. The user supplied group_index inside of the `adapt_zetas`
function was ignored when retrieving current weights. This has now been
fixed, and a test has been added of the use case that originally
captured the broken behavior, and maintain correctness in the future.

This error might have gone undiscovered in the past due to SAMS not
frequently being used on a residue other than the first titration group.
Typical past use case was SAMS on one ligand/amino acid in solvent.

This change addresses the need by:

* Fix group_index in SelfAdjustedMixtureSampling.adapt_zetas call to
`get_gk`.
* Add test where SAMS calibration is performed on a residue other than
the first residue in the system.