Large overhaul and cleanup
Description:

This commit contains a whole lot of changes. During cleanup and
reorganization, it became useful to fix a few bugs here and there.
Unfortunately the diffs aren't very useful because a lot of files
were shuffled around. I'm summarizing the changes here to the best of my
abilities.

Major updates:
* protons now has an `app` layer style. The user can easily set up
constant pH simulations using protons.app.ForceField, protons.app.Modeller,
and protons.app.Topology. These objects can easily access the force field
and residue definitions used for constant-pH simulation.
    * Modeller automatically picks constant pH variants, including `AS4`, `GL4`, `HIP`.
    * Forcefield knows the location of `amber10-constph.xml`, `gaff.xml`,`gaff2.xml` and ion parameters.
    * Topology loads bond definitons from  `bonds-amber10-constph.xml` when using pdbfiles.

* ConstantPHSimulation, and ConstantPHCalibration classes have been added to
 facilitate easy simulations, automatically attaching contexts, book keeping and reporting.
 These are analogous to openmm.app.Simulation.
* Residue state updates can now be selected using an external object type
known as a StateProposal. Several basic types have been included. This
is of course extensible to more clever update schemes.
* ProtonDrive constructors have been simplified as much as possible. They have less arguments
and some functionality has been moved to separate functions.
* ProtonDrives now have an attach_context function. This allows the user
to provide the context that will be simulated, and the integrator that
is to be used. It removes the need for context bookkeeping. The integrator
and context are stored internally so that the user does not need to pass
them to any function.

Deprecated:
* The old calibration API has been removed, since it is now completely deprecated.
It is no longer used, and the recommended way to run calibrations is to use app.ConstantPHCalibration.
* The old ff module with link to forcefield files is no longer used, in favor of the new app structure.

Minor updates:
* For efficiency and ease of accessing, the ProtonDrive objects only keep track of the last update attempt.
* netCDF formats have been updated to match the new drive internal data structure
* Doc strings have been cleaned up.
* Forcefield files are now located under protons/app/data/
* Forcefield files have been renamed from `*-protons.xml` to `*-amber10-constph.xml`, in order to
emphasize origin of the parameters.
* Forcefield files use canonical atom names from `pdbNames.xml` for better compatibility with
pdb files and modeller.
* Propagations_per_step argument actually used inside of NCMC protocol.
* GHMC integrator has new properties, for compatibility with the GBAOABIntegrators
* Setup structure updated to track new directories
* Add pytest to setup.py
* Update python versions for travis.

Bugfixes:
* Major bug in hydrogen definition files was fixed. Histidine residues would automatically
* Major indexing bugs in ProtonDrive._attempt_state_change were fixed. Previously,
 it was incorrectly assessed whether the titration state would be changed, and ncmc protocols were skipped.
* Removal of temp_files bug in ligands xml generation was fixed.
* Bug in NME atom names in the forcefield definitions was fixed by using pdbNames.xml.