Attempt to avoid segfaults on travis
Why:

It seems that netCDF related tests have a tendency to cause segmentation
faults, only on travis. To avoid this from causing travis failures,
 these tests will only be available for running locally.

It is possible that it is not netCDF related. Because of the difficulty
reproducing this outside of travis, it is hard to pinpoint where the exception
is coming from.

For now, avoiding the tests on travis and running them locally
 seems the best solution.