Add pH correction feature
Why:

The internals of the code require all pH corrections to be applied by
the user beforehand by specifying a corrected g_k.

The new methodology cleans up and reuses parts of the code that were
responsible for pH corrections in the old calibration API. The new
convention is that all g_k values specified are for equal sampling of all
states, and for residues besides amino acids, the desired log populations
must be specified for a given pH in the ffxml files with protons residues.
These residues use <Condition/> tags which are placed inside each <State/>
block. For an example see the imidazole explicit test system, under
 protons-imidazole-ph-feature.xml.