Add serializer for states, residues inside the drive
Why:

To reset/continue a simulation, we need to store the internal state of
the proton drive.

The first step is to serialize the titration groups, and the internal state
information.

Note:

Implementing this revealed a bug in the pH correction feature.
This bug was fixed here to allow the code to run.