Add framework for deserialization _TitratableResidue
Why:

Python does not support multiple constructors. I've reshaped init to be
barebones. The proper way to instantiate the new class is by using the
 `from_lists`, or `from_serialized_xml`, methods.  Methods names as such
 are a common paradigm in other python libraries, such as datetime or pandas.

The inner workings of deserializing _TitrationState are not done yet, and
 will need to be completed first before this is operational.

 We should also write tests that check that the serialization and deserialization
 don't lead to a loss of information, though some new unit test.

Note that the beautifulsoup4 library has been added as a dependency as a
robust way to parse serialized pandas dataframes as html tables.