Add proposal mechanism for selecting neutralizing ion swaps.
Why:

The proposals.py module has been extended with a new proposal mechanism
that interfaces with saltswap. In order to maintain charge neutrality,
the proposal mechanism uses saltswap functionality to change ions into
water and vice versa, with the objective to counter any charge change
from changing between protonation states.

The output from the proposal function should contain all the information
 necessary to interact with the saltswap.Swapper.update_fractional_ion
 method, as well as the proposal probability that needs to be included in
 the Metropolis-Hastings acceptance probability.

What has been done:

* A _SaltSwapProposal baseclass has been added that defines the interface
of proposals for coupling protons with saltswap. This interface uses the
propose_swaps function, which takes in the ProtonDrive, and the charge
that needs to be countered with ion swaps. As needed for calling
Swapper.update_fractional_ion, it returns the indices of the ion/water
molecule in the Swapper.stateVector, and the initial and final
parameters of the ion/water. It also returns the log ratio of the
reverse/forward proposal probabilities, which is to be added to the
log acceptance ratio for the Metropolis-Hastings/NCMC acceptance
criterion.

* The UniformSwapProposal class has been implemented. It uniformly selects
from the pool of waters, cations, or anions, and calculates proposal
probability. It automatically works out the minimum set of swaps in
waters, and should prevent accumulation of ions.