Set excess_ions from attach_swapper
Why:

As suggested by Greg, to get a better idea of the available ion pool it
makes sense to use the total counter charge ions available in the
swapper at the start of the simulation, rather than starting at 0. This
addition to the attach_swapper method retrieves the right number.

What:

* drive.attach_swapper method uses the identity counts from Swapper
to calculate what the starting excess ions should be.