Extract Epik calculation
Why:
This refactor makes it easier to run epik outside of the protons framework
and then import it into protons after the fact.

In the mean time, users can fix issues with the atom names, or even
use tools other than Epik to generate the input file for ffxml
generation.