Implement Chen-Roux ion scheme
Why:
This implements the ion selection scheme as mentioned in work by Chen
and Roux 2015. It replaces the old scheme, which was found to potentially
accumulate ions when dealing with charge differences larger than single
units. We can revisit the scheme at a later date to see if the problem
 can be addressed. In the mean time, we will be using the Chen-Roux
 scheme. In a subsequent commit, a test will be added to ensure that
this can be caught more easily for future methods.

What:

* The UniformSwapProposal has been replaced by the Chen-Roux proposal.

Signed-off-by: Bas Rustenburg <bas.rustenburg@choderalab.org>