Add test for accumulation of ions for a variety of residue types
Why:

This adds a bunch of test cases, and a more robust way of telling
whether the implementation of the ion-swapping could lead to
unintended accumulation of ions.

The way it works, it checks whether the number ions that have been added (or removed)
over a long trajectory is larger than the maximum span of the residue
charges. If more are added (or removed) than the span, ions are
accumulating (or depleting) over multiple iterations.

What:

A new test class called TestIonSwap has been added. It is used to
test the Chen-Roux 2015 method of keeping the direction of alchemical
charges the same.

Signed-off-by: Bas Rustenburg <bas.rustenburg@choderalab.org>