Remove outdated test
Why:

This test was replaced by a more effective and faster test. The old test
actually only covered one residue, and required the instantiation and
simulation of an actual system. New tests in TestIonSwapping can actually
test the accumulation independently, without the need for Cuda and even
OpenMM to make sure the algorithm is correct.

 What:

* This removes the old accumulation test.

Signed-off-by: Bas Rustenburg <bas.rustenburg@choderalab.org>