Add calibration system creating functionality.
Why:

To calibrate, a solvent system needs to be added. This adds a feature
that extracts a ligand from a pdb file (that potentially contains the
protein and solvent that needs to be removed).

We then need to add adds hydrogens and solvent. In the future, this
function could be expanded to set up a system with a protein in a more
straightforward way as well.

Tests still need to be implemented for these methods.

What:

* functions have been added to extract a ligand and set up a calibration
system.

Signed-off-by: Bas Rustenburg <bas.rustenburg@choderalab.org>