Add skeleton class for bookkeeping titration attempts
Why:

* Certain information needs to be recorded for analysis purposes.
Currently, this is being done in a tuple with a predetermined structure.
The number of items we want to extract has increased, and it's
more sustainable to have a clean container class with properties
in order to pass data around without accidental overwriting.

This change addresses the need by:

* Adding the _TitrationAttemptData class to the driver module.