Prevent non-integer charges from causing infinite while loop
Why:

The saltswap ion proposal in proposals.py includes a while
loop to resolve charges and select ions. However,
this while loop involved comparing if an integer was greater than 0.
The type of the titrationstate total charge returned was a float,
so this in some cases leads to infinite loops because of a slight
float remainder.

This should adress that error.

What:

* Titrationstate total_charge function returns rounded integers
instead of floats.
* titratableresidue total_charges should now return lists of integers.