Add detection for while loop in future
Why:

This loop can get stuck in an infinite loop at times when
charges don't add up right. To catch this in the future,
I added a check to just raise an error if this loop is repeated
a very large amount of times (1000), and prevent an infinite
runtime.

What:

* add clause to raise Runtime error in salt proposal
if swaps are not resolved in 1000 tries.