Make a deepcopy of saltVector arrays
Why:

This is a bug that is potentially responsible for
the accumulation of ions we have observed in
ion swapping simulations. In short:
using slice notations ( listname[:]) returns
a copy for python lists, but for ndarrays, it
returns a reference. We should explicitly make a deep copy instead.

This change addresses the need by:

*  Make copies of a swapper.stateVector in
NCMCProtonDrive.attempt_state_change