Add proposal class that can flip carboxylic acid
Why
---

In order to speed up conformational sampling of carboxylic acids
it is possible to relocate the hydrogen dummy while it is
deprotonated. This should come at little energetic cost.

What
----

A class was added to proposals.py called "COOHDummyMover" which
can take atom indices for a carboxylic acid moiety, and flip
along bonds.