Add COOH patching for residue templates
Why
---

This adds a feature to patch residue templates that have carboxylic acid
groups in them. Protons parametrization sometimes assigns o instead of
oh types to the hydroxy oxygen, leading to the lack of an angle term
between the C-O-H of carboxylic acids. Additionally, using just oh for the
one oxygen would lead to asymmetry in the deprotonated state.

Additionally, this patch introduces COOH declarations into the XML,
which tell the protondrive to add dummy movers to the system. These
are presumed to greatly enhance the configurational sampling of the
hydrogen position in the deprotonated state.

What
----

- A new module called template_patches was added under protons.app.
 it contains a function called patch_cooh, which patches residues
 containing titratable COOH groups.
 - A test was added to test_cooh.py to ensure it patches viologen
 correctly.

Signed-off-by: Bas Rustenburg <bas.rustenburg@choderalab.org>