Add special carboxylate protein residues
Why
---

The forcefield generating scripts were updated to make new carboxylic
acid residues (GLH and ASH), which only have one hydrogen. These are
meant to replace the GL4 and AS4 residues. The 4 hydrogen residues were
originally defined to get around the torsional barriers for hydrogens,
but the script was now updated to patch these residues with a COOH
dummy mover. This allows speeding up the sampling with a torsional
barrier.

What
----
- 2-process_raw_xml_file.py was updated to replace ASH and GLH residues with titratable ones.
- The new titratable residues have COOH statements.

Signed-off-by: Bas Rustenburg <bas.rustenburg@choderalab.org>