Add SAMS weight table
Why
---

SAMS can be used for more than just calibrating a single residue at a
time. Every combination of residue titration states could be seen as a
single state in the free energy calculation. For relatively small
number of residues, this could allow a pretty efficient sampling of
protonation states in order to estimate free energy differences.

What
----

- A class called "WeightTable" was added to calibration.py
- An enum class was added called SAMSApproach, to be used to enumerate
the choices for SAMS calculation types.
- A file called test_calibration.py was added to test. It contains
tests of the table/enums .

Signed-off-by: Bas Rustenburg <bas.rustenburg@choderalab.org>