Fix COOH sampling
Why:
After re-adding missing forcefield terms, it became clear that the
implementation of the COOHDummyMover did not respect all angle terms,
and was unfavorable for dihedrals. I implemented a new strategy that
uses a series of grid points by using internal coordinates.

Using importance sampling with normalized potential energies as weights,
(and the assumption that the dummy has no non-bonded energy, which is
correct for amber10 and gaff) we have a rejectionless method that can
resample the hydrogen in any position on any of the two oxygens.

What:
- COOHDummyMover internals were redesigned.
- Driver was modified to only resample velocities for atoms with
perturbed coordinates.
- Tests were updated for the COOHDummymover

Signed-off-by: Bas Rustenburg <bas.rustenburg@choderalab.org>