Add run script and bugfixes to facilitate run
Description
-----------

This adds the calibration-v1.py script, and makes a few changes to
the prepare_run_from_ffxml script and its dependencies. The example
input is added under the name "run_calibration-v1.py"

Changes
-------

- Saltswap now also stores the salt concentration, and serialization
of the saltwap information depends on the Salinator instead of the
Swapper class.
- The input pdbx file is now stored in the checkpoint file as a means
of containing the correct topology within the checkpoint.
- Convenience functions for deserializing and serializing have been added
to protons.scripts.utilities.

Bugfixes
--------

- protons.app.driver._TitrationState serialization was fixed for cases where more than
one force was stored in the _forces object
- Just in case more than one proton drive is present in an xml file,
the protondrive deserialization function now uses the SAMSState associated
with the selected (first) protondrive in the file, rather than any
(and therefore possibly multiple) SAMSState tags

Signed-off-by: Bas Rustenburg <bas.rustenburg@choderalab.org>