Add option to calculate bar for all pairs, not just pairs with first state
Why
---

For most purposes in the current state of the module the bar energies
between state 1 and others are sufficient but for debugging and extended
analysis efforts the bar estimates between all other states are helpful.

What
----

This adds a boolean toggle to include other pairs in the output for the
bar_all_states function in analysis.py.

Signed-off-by: Bas Rustenburg <bas.rustenburg@choderalab.org>