Expand/correct run scripts for calibration
Why
---

In the process of setting up runs, several changes were needed to meet
the needs in production. Largest change is the incorporation of the
work_interval flag to the calibration script, for the ncmc reporter.
This flag lets the user specify how often to store intermediate work
values for NCMC protocols. This feature is useful for benchmarking and
debugging.

Signed-off-by: Bas Rustenburg <bas.rustenburg@choderalab.org>