Implement the 4-stage scheme in ConstantPHSimulation and NCMCProtonDrive
Why
---

This allows for using the new 4 stage sams scheme from the familiar
ConstantPH Simulation API.

What
----
The parameters min_burn, min_slow, and min_fast now control the number
of iterations required to switch between each scheme and avoid early
convergence. The burn-in phase helps with initial guesses that are far
off, but afterwards it goes to a regular SAMS run. The flatness criterion
still determines when to switch from fast to slow, and additionally
1/10 * flatness_criterion kicks in equilibrium sampling.

PS:
To use the old scheme, set the burn-in period to 0, and set the
min_fast period to a very large number of iterations.