Bugfix: COOH move detection from xml now works with and without moves
Why
---
I changed the iteration structure so that it works both when there are,
as well as aren't any moves present in the MCMoves block. This has to do
with the StringElement class in lxml showing unusual iteration behavior.

I also made the tests of these features use less iterations so that they
are cheaper on computers without GPU and can still catch logical bugs.