Update calibration to use new deviation and counts
This updates the SAMSCalibrationEngine to use the new way of calculating
the deviation from the target histogram. It also adds a way to count the
states visited in the equilibrium phase using the same mechanism as is
done during the rest of the calibration.

The simulation class was extended to make use of this new feature, and
it now also resets the counts at the start of every new stage.