Add NCMC benchmark dataframe functionality
Why
---
To analyze NCMC benchmark datasets efficiently, I've added a function
that can generate a pandas dataframe with relevant labels and variables
in it. This can be used subsequently to easily analyze statistics and
generate plots. For instance, seaborn has very good support for using
a pandas dataframe with its built in functions.

Additional change
-----------------
The stitch calibration and dataset to array functions were modified to
also return the SAMS stage, since this is useful for filtering out
equilibrium samples from a dataset.