Expose scripts through a simple interface
Why
---
For easy use of the basic scripts, I've added a command line interface
that simply links the previously prepared scripts through a shortcut.

For instance, the parameterization script can now be run from anywhere
as

`protons param inputfile.json`

or, running a simulation

`protons run inputfile.json`
.

This is easier to maintain than having separate scripts to copy, and
version control.

What
----

I've added a cli.py file and entry points in setup.py that point to it.
This now exposes the `protons` command line script on people's path when
installing protons.