Include formal charges in parameterization
This internalizes some of the available openmoltools features that
parameterize force fields for ligands using gaff. I've taken the
openeye charge assignment function and added an option to assign
formal charges to molecules. This should help parameterizing ligands
that do not have proper formal charge information in the input mol2.