Add SMILES option to param script
This adds the option to start a ligand parameterization run from a
SMILES string instead of a maestro file. It adds a new feature to the
ligands model that uses Schrodingers smiles_to_mae utility so that you
can easily start an Epik run from smiles.

It also cleans/updates the cli examples.