Add option for BAR on only equilibrium samples
This adds a flag to the bar_all_states function to easily filter out and
only analyze data that has reached the equilibrium stage.

I've also added an option to return nan for transitions where forward or
reverse trajectories aren't present for an individual pair of states.
Previously this would result in an error from the BAR code that was hard
to interpret.