Fix the title of molecules used for charges
The molecule title should not have spaces in it. This causes problems
where antechamber silently replaces all charges with 0.0. This fix adds
a unique title without spaces for the purpose of charge generation.
It does not affect the final title of the molecule in the template, but
works around the issue.