Add option for conformers to parametrization
This adds an option to set the omega max conformers for ligand
parametrization. To set max conformers, use the "omega_max_confs" flag
in the input. To use Dense conformers (aka as many conformers as it
takes) use -1. This can help with ligands that have convergence issues
during ELF10 charge generation. For information on the Dense options
you can also see:

https://docs.eyesopen.com/toolkits/python/omegatk/OEConfGenConstants/OEOmegaSampling.html#OEConfGen::OEOmegaSampling