Use YAML parser for JSON input
Since YAML is a superset of JSON, this does not break existing features,
but may allow for easier comments and hand editing of files.

Considering future plans. To minify files, we could always convert YAML
files to JSON before transmitting, and we should be able to cover
both file types with a YAML schema.