Initial abstraction of ion selection
This implements a new ions submodule that will feature utilities for
determining the treatment of ions in concordence with protonation state
effects.

Rationale: In the current implementation, the ions inserted/removed
are determined on the fly by a MC proposal class in
protons/app/proposals.py. For robustness, it is better to define this as
part of the definition of the state of the residue.

Example: rather than proposing to protonate aspartate, and then propose
 to change a water molecule into Cl- to maintain charge neutrality, we
 can a priori associate the protonated form of aspartate to be
 ASH:Na+Cl-, and the deprotonated form would be ASP-:Na+:H2O.
 Where this matters is when we set the protonation state at the start of
 the simulation in two different simulations, where it is possible that
 to define the states as ASP-:Na+:H2O and ASH:2 H2O for neutrality.
 While both are valid, we should have a canonical form. The new
 abstraction allows specification of what form should be canonical.

 Changes in this commit:
 The new ions module has functions for selecting ions based on one
 reference state charge (should by definition be the state with index 0
 in the residue template) and a given protonation state charge. Which
 ions are selected depends on a rule. This commit contains 3 rules:
 - No ions : neutralizing background charge, part of OpenMMs PME
 implementation is used for charge neutrality.
 - Counter ions: Changing from neutral to cation is accompanied by an
anion being added to solvent, and removing a cation ion removes an anion
, and similarly for anions. A pair of opposing charges is always changed
on or off at the same time while maintaining neutrality.
 - Replacement ions: Changing a cationic protonation state to neutral
will add a replacing cation in solvent, and changing a neutral state to
positive will remove a cation in solvent, and similarly for anions.
A pair of charges with the same sign are always changed at the same time,
phasing one in, and one out, maintaining overall neutrality.

The _TitrationState defintion in driver.py now contains information
about which ions accompany each state. This is initially set to 0 for
each state. The NCMCProtonDrive.attach_swapper function has been renamed
to `enable_neutralizing_ions`, and allows specification of one of the
three rules above, which then updates the ions specified in
the _TitrationState object. Calls to attach_swapper in other modules
have been renamed as well.

Added as well are three tests of the three rules to the test_explicit
module.

Still remaining is updating the Proposal mechanism, this commit only
adds the bookkeeping and specification of ions, not the actual changing.