Refactor updating of ions based on new strategy
This uses the previously implemented mechanism for determining the ions
needed alongside each titration state to update ionic states with
saltswap.

The refacting is quite extensive, as I've included a few refactors to
reduce code duplication and make certain parts of the process more
modular. This includes things such as the way the state vector is
updated, which is now done primarily by a function in the new
saltswap_utils module.

I've updated and exposed the
NCMCProtonDrive.set_titration_state function to allow it to be used with
ions. In most cases where it was previously used, we still choose ions
externally, and update those separately, but now one can also change the
protonation state without using MCMC (for instance, to sample
equilibrium configurations) with the right ions alongside it.

The SaltSwapProposal class no longer selects how many ions should be
added/deleted, since this is now dependent on the state, rather than
being derived on the fly. I've simplified them and I removed the
old OneDirectionChargeProposal class with a UniformSwapProposal class
that selects ions/waters to swap based on a delta_ion count.
This also fixes a bug in the calculation of the number of possible
ions to pick.