Implement (annealed) importance sampling
This adds (annealed) importance sampling as a method. Essentially, all
that is changed is that proposals are never accepted. The ProtonDrive
classes are extended with a new __init__ argument that determines
whether MCMC or importance sampling is to be used.

I've added two tests that make sure that importance sampling is working
as expected.

I still need to update scripts to be able to set up importance sampling.