Add importance sampling options to prep scripts
This adds the possibility to set up a system for use with importance
sampling. Two scenarios are possible with the new setting:

Option 1:
The user specifies a specific protonation state for each residue, and
one checkpoint is set up with the residues in said state with matching
counter ions. (Still need to add the counter ions as option in script.)
To do this one adds the "importance" block  with the list of states for
each residue (as integer) under "titration_states".

Option 2:
The user specifies "importance" : {"systematic": true}. The script will
generate a run checkpoint for every possible state.