Update script ion features
This adds the new neutralizing ion implementation and merges it with the
importance sampling methods to scripts.

Notable change to mention:
previously, ions were specified in the run script, not the prepare
script. This is now (more intuitively) part of the prepare script.

By default, saltswap is used but it only adds ions depending on the
charge rule. Previously, the run script would add ions per individual
state depending on the counter-ion flag in the run script. Now, since
ions for each state are pre-enumerated in the prep stage, there is no
ambiguity and less room for error.