Switch script input to toml
This switches the input file format for the protons scripts to toml

Since these files are likely to be user constructed it makes sense for
clear denominations to exist between the different keys in the file.

TOML provides a balance between the many delimiters in JSON files and
the succinctness of YAML.

Bonus: log messages in run_prep_ffxml are now decorated and expanded.