Check in new analysis modules
This is the first check-in of the new analysis and data processing
modules. The old code has been replaced/redesigned in an attempt to
clean up, streamline and future-proof the analysis of constant-pH
simulations.

The new structure of the module is as follows
protons/app/analysis has three submodules
- `plotting`, for useful shortcuts to complicated plots
- `processing`, for converting raw data (nc files) into malleable
pandas dataframes
- `statistics`, for performing numerical data analysis on the dataframes
produces, such as BAR, et cetera.

Still necessary to add:
- tests for plots
- tests for other data types (such as multi residue)
- tests for plots
- conversion of a few old plotting functions such as heatmaps