Keep track if bias in dataframes
Just because a run isn't a calibration, doesn't mean that there is no
bias. In the case of equilibrium runs, the g_k value is set by the user
and is constant. This adds a lookup from the metadata information in
the netCDF file and adds new columns to the pandas dataframe containing
this info.